+++ lib/status/Cargo.toml
@@ -18,7 +18,3 @@ tracing.workspace = true
 
 reth-tasks.workspace = true
 zksync_os_raft.workspace = true
-
-[dev-dependencies]
-reqwest = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
