+++ lib/status/src/lib.rs
@@ -53,45 +53,3 @@ pub async fn run_status_server_on_listener(
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use reth_tasks::{RuntimeBuilder, RuntimeConfig, TokioConfig};
-    use std::net::{Ipv4Addr, SocketAddrV4};
-    use tokio::net::TcpListener;
-    use tokio::runtime::Handle;
-
-    #[tokio::test]
-    async fn status_server_on_listener_responds() {
-        let listener = TcpListener::bind(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0))
-            .await
-            .unwrap();
-        let port = listener.local_addr().unwrap().port();
-        let runtime = RuntimeBuilder::new(
-            RuntimeConfig::default().with_tokio(TokioConfig::existing_handle(Handle::current())),
-        )
-        .build()
-        .unwrap();
-        runtime.spawn_critical_with_graceful_shutdown_signal(
-            "status server",
-            |shutdown| async move {
-                run_status_server_on_listener(listener, shutdown, None)
-                    .await
-                    .unwrap();
-            },
-        );
-        let deadline = tokio::time::Instant::now() + std::time::Duration::from_secs(5);
-        let resp = loop {
-            match reqwest::get(format!("http://localhost:{port}/status")).await {
-                Ok(resp) => break resp,
-                Err(err) if tokio::time::Instant::now() < deadline => {
-                    tracing::debug!(%err, "status server not ready yet");
-                    tokio::time::sleep(std::time::Duration::from_millis(25)).await;
-                }
-                Err(err) => panic!("status server did not become ready: {err}"),
-            }
-        };
-        assert_eq!(resp.status(), 200);
-    }
-}
